+++ 2DCollisions/Init.cpp
@@ -104,7 +104,7 @@ GLvoid ChangeToFullScreen()
 /////
 ///////////////////////////////// CREATE MY WINDOW \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*
 
-GLuint CreateMyWindow(LPSTR strWindowName, GLint width, GLint height, DWORD dwStyle, GLboolean bFullScreen, GLvoid* hInstance)
+GLuint CreateMyWindow(LPSTR strWindowName, GLint width, GLint height, unsigned long dwStyle, GLboolean bFullScreen, GLvoid* hInstance)
 {
 	GLuint hWnd;
 	WNDCLASS wndclass;
